+++ app/Filament/Resources/DocumentSequences/Pages/EditDocumentSequence.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Modules\ERP\Filament\Resources\DocumentSequences\Pages;
 
 use Filament\Resources\Pages\EditRecord;
+use Modules\ERP\Filament\Resources\DocumentSequences\Actions\DocumentSequenceActions;
 use Modules\ERP\Filament\Resources\DocumentSequences\DocumentSequenceResource;
 use Override;
 
@@ -12,4 +13,15 @@ final class EditDocumentSequence extends EditRecord
 {
     #[Override]
     protected static string $resource = DocumentSequenceResource::class;
+
+    /**
+     * @return array<int, \Filament\Actions\Action>
+     */
+    #[Override]
+    protected function getHeaderActions(): array
+    {
+        return [
+            DocumentSequenceActions::reset(),
+        ];
+    }
 }